+++ EInvoice.Infrastructure/Migrations/EInvoiceContextModelSnapshot.cs
@@ -24,11 +24,11 @@ namespace EInvoice.Infrastructure.Migrations
 
             modelBuilder.Entity("EInvoice.Domain.Entities.Client", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
 
                     b.Property<string>("Address")
                         .IsRequired()
@@ -62,14 +62,14 @@ namespace EInvoice.Infrastructure.Migrations
 
             modelBuilder.Entity("EInvoice.Domain.Entities.Invoice", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
 
-                    b.Property<int>("BuyerId")
-                        .HasColumnType("int");
+                    b.Property<long>("BuyerId")
+                        .HasColumnType("bigint");
 
                     b.Property<DateTime>("InvoiceDate")
                         .HasColumnType("datetime2");
@@ -89,8 +89,8 @@ namespace EInvoice.Infrastructure.Migrations
                         .HasMaxLength(50)
                         .HasColumnType("nvarchar(50)");
 
-                    b.Property<int>("SellerId")
-                        .HasColumnType("int");
+                    b.Property<long>("SellerId")
+                        .HasColumnType("bigint");
 
                     b.HasKey("Id");
 
@@ -103,11 +103,11 @@ namespace EInvoice.Infrastructure.Migrations
 
             modelBuilder.Entity("EInvoice.Domain.Entities.InvoiceItem", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
 
                     b.Property<double>("Discount")
                         .HasColumnType("float");
@@ -129,8 +129,8 @@ namespace EInvoice.Infrastructure.Migrations
                         .HasMaxLength(20)
                         .HasColumnType("nvarchar(20)");
 
-                    b.Property<int>("InvoiceId")
-                        .HasColumnType("int");
+                    b.Property<long>("InvoiceId")
+                        .HasColumnType("bigint");
 
                     b.Property<string>("ProductDescription")
                         .IsRequired()
@@ -186,11 +186,11 @@ namespace EInvoice.Infrastructure.Migrations
 
             modelBuilder.Entity("EInvoice.Domain.Entities.Organization", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
 
                     b.Property<string>("Address")
                         .IsRequired()
